add sparkling wines parser support global project restructuring small naming issues fixes

+++ src/main/java/com/wine/to/up/winestyle/parser/service/dto/WineDto.java
@@ -19,10 +19,10 @@ public class WineDto implements Serializable {
     private String color;
     private String country;
     private String region;
-    private Double volume;
+    private Float volume;
     private String strength;
     private String sugar;
-    private BigDecimal price;
+    private Float price;
     private String grape;
     private String tastingNotes;
     private Double rating;
